Android <-> Database Connection Tutorial
This is Part I of a project to build a full stack
application that bridges surveyors and surveyees
via application interfaces linked to a database

It follows a tutorial (url below) to create an
android application and a MySql DB (MariaDB) via
a PHP connective application, utilizing the XAMPP
software. Although I have experience with SQL/DB
administration and Android applications separately,
my goal is to bridge and grow my knowledge in how
best to connect them - Particularly for early/small
proof-of-concept development.

I plan to do a lot of testing and tinkering to
better understand the infrastructure I'll need
to develop and test the Transcription mobile app

https://www.androidhive.info/2012/05/how-to-connect-android-with-php-mysql/
Although a little old, it gave me the opportunity to
rapidly get a base to start with, and begin updating
it. Thanks to Ravi Tamada of AndroidHive